requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
Pillow>=10.0
django-filter>=23.0
psycopg2-binary>=2.9.9
